market-server/requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
boto3==1.10.20
botocore==1.13.20
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
decorator==4.4.1
Django==2.2.7
django-braces==1.13.0
django-cors-headers==3.2.0
django-crontab==0.7.1
django-oauth-toolkit==1.2.0
django-storages==1.7.2
django-user-agents==0.4.0
djangorestframework==3.10.3
django-rest-swagger==2.1.2
drf-yasg==1.17.1
docutils==0.15.2
firebase-admin==4.0.0
flex==6.14.1
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.23.1
google-cloud-core==1.0.3
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
httplib2==0.14.0
idna==2.8
idna-ssl==1.1.0
imageio==2.6.1
imageio-ffmpeg==0.3.0
itunes-iap==2.5.0
jmespath==0.9.4
moviepy==1.0.1
msgpack==0.6.2
multidict==4.5.2
mysqlclient==1.4.5
numpy==1.17.4
oauth2client==4.1.3
oauthlib==3.1.0
pandas==0.25.3
phonenumbers==8.11.1
Pillow==6.2.1
prettyexc==0.6.0
proglog==0.1.9
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycares==3.0.0
pycparser==2.19
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
pyproj==2.6.1.post1
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
tqdm==4.38.0
typing==3.7.4.1
typing-extensions==3.7.4.1
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.25.7
user-agents==2.0
uWSGI==2.0.18
xlrd==1.2.0
XlsxWriter==1.2.6
xmltodict==0.12.0
yarl==1.3.0